port to Linux (now Ubuntu 13.10 server)

+++ src/conf_private.h
@@ -9,7 +9,7 @@ extern "C" {
 
 struct _read_ctx {
     const char *buf;
-    int pos;
+    off_t pos;
     int sz;
 };
